requirements.txt
aiosmtplib==2.0.2
annotated-types==0.6.0
anyio==3.7.1
bcrypt==4.0.1
blinker==1.7.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
coverage==7.3.2
cryptography==41.0.5
dnspython==2.4.2
email-validator==2.1.0.post1
fastapi==0.104.1
fastapi-mail==1.4.1
greenlet==3.0.1
h11==0.14.0
httpcore==1.0.2
httpx==0.25.1
idna==3.4
imageio==2.33.1
iniconfig==2.0.0
Jinja2==3.1.2
lazy_loader==0.3
MarkupSafe==2.1.3
networkx==3.2.1
NudeNet==2.0.6
numpy==1.26.3
opencv-python-headless==4.9.0.80
packaging==23.2
passlib==1.7.4
pillow==10.2.0
pluggy==1.3.0
progressbar2==4.3.2
psycopg2-binary==2.9.9
psycopg2-pool==1.1
pycparser==2.21
pydantic==2.4.2
pydantic-settings==2.0.3
pydantic_core==2.10.1
pydload==1.0.9
PyJWT==2.8.0
PyMySQL==1.1.0
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0
python-dotenv==1.0.0
python-multipart==0.0.6
python-utils==3.8.1
requests==2.31.0
scikit-image==0.22.0
scipy==1.11.4
sniffio==1.3.0
SQLAlchemy==2.0.23
starlette==0.27.0
tifffile==2023.12.9
typing_extensions==4.8.0
urllib3==2.1.0
uvicorn==0.24.0.post1
